53/r/chatmessage (#63)
* refactor: 불필요한 코드 제거 & JsonIgnore

* refactor: message가 chatMessagesDto에 담겨서 return되도록 수정

* feat: 채팅방 목록 기능 수정 & 불필요한 코드 제거

* feat: chatRoomMember의 lastViewMessageId를 업데이트하는 기능 추가

+++ src/main/java/com/example/be8arm/domain/chat/chatRoom/dto/ChatRoomDetailDto.java
@@ -20,10 +20,10 @@ public class ChatRoomDetailDto {
 	private String lastMessageContent;
 	private String lastMessageSender;
 	private LocalDateTime lastMessageDate;
-	private int unreadMessagesCount;
+	private Long unreadMessagesCount;
 
 	public ChatRoomDetailDto(Long chatRoomId, String chatRoomName, String imgUrl, String lastMessageContent,
-		String lastMessageSender, LocalDateTime lastMessageDate, int unreadMessagesCount) {
+		String lastMessageSender, LocalDateTime lastMessageDate, Long unreadMessagesCount) {
 		this.chatRoomId = chatRoomId;
 		this.chatRoomName = chatRoomName;
 		this.imgUrl = imgUrl;
